Mention the repository installer (again, yes)
I was looking into installing (only) Jellycon but I couldn't figure out which repository to use. Then as a fallback I used `Jellyfin for Kodi` and voila Jellycon was right there. I know, stupid me but this 1 step for dummies like me might help. Even though it clearly mentions that the repo "allows for easy install of all our add-ons".

I'm sure there'll be a better way of documenting this so I'm welcome to that as well if this isn't good or more confusing.